ReceiptsDetailViewController is a subclass of QLPreviewController
